Final Hand segmented with erosion.
Added support for detecting no motion. If no motion, it instructs the
user to Move His Hand.
So, Final segmentation is done. We might need some tweaks here and there
based on the application. We can even make the constants dynamic in near
future. For now lets park this as an individual module. We'll combine
the modules as and when needed.